# app/schemas/charity_project.py
from datetime import datetime
from typing import Optional

from pydantic import BaseModel, ConfigDict, Field, PositiveInt


class CharityProjectBase(BaseModel):
    name: str = Field(
        ...,
        min_length=5,
        max_length=100,
    )
    description: str = Field(
        ...,
        min_length=10,
    )

    model_config = ConfigDict(extra='forbid')


class CharityProjectCreate(CharityProjectBase):
    full_amount: PositiveInt


class CharityProjectUpdate(BaseModel):
    name: Optional[str] = Field(
        None,
        min_length=5,
        max_length=100,
    )
    description: Optional[str] = Field(
        None,
        min_length=10,
    )
    full_amount: Optional[PositiveInt] = None

    model_config = ConfigDict(extra='forbid')


class CharityProjectDB(CharityProjectCreate):
    id: int
    invested_amount: int
    fully_invested: bool
    create_date: datetime
    close_date: Optional[datetime] = None

    model_config = ConfigDict(
        from_attributes=True,
        extra='forbid',
    )